[ADD] #7 SQLite/Room 설정, Dao/Entity 생성, Query/Coroutine작성
Former-commit-id: 2230226b137a823342485e216e992f8d13cb0a63

+++ Smile_Android/app/src/main/java/com/ssafy/smile/domain/repository/PortfolioRepository.kt
@@ -7,8 +7,8 @@ interface PortfolioRepository {
     suspend fun getPortfolio(photographerId: Long)
     suspend fun getPosts(photographerId: Long)
     suspend fun uploadPost(images: MutableMap<String, RequestBody>)
-    suspend fun uploadPost(latitude: Float, longitude: Float,
+    suspend fun uploadPost(latitude: Double, longitude: Double,
                            detailAddress: String, category: String, images : List<MultipartBody.Part>)
-    suspend fun uploadPost(latitude: Float, longitude: Float,
+    suspend fun uploadPost(latitude: Double, longitude: Double,
                            detailAddress: String, category: String, images : HashMap<String, List<MultipartBody.Part>>)
 }
